Move fishing rod selection code into AutoFishRodSelector
