+++ src/main/java/net/wurstclient/hacks/AutoFishHack.java
@@ -9,9 +9,6 @@ package net.wurstclient.hacks;
 
 import net.minecraft.client.network.ClientPlayerEntity;
 import net.minecraft.client.util.math.MatrixStack;
-import net.minecraft.enchantment.EnchantmentHelper;
-import net.minecraft.enchantment.Enchantments;
-import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.entity.projectile.FishingBobberEntity;
 import net.minecraft.item.FishingRodItem;
 import net.minecraft.item.ItemStack;
@@ -24,6 +21,7 @@ import net.wurstclient.events.RenderListener;
 import net.wurstclient.events.UpdateListener;
 import net.wurstclient.hack.Hack;
 import net.wurstclient.hacks.autofish.AutoFishDebugDraw;
+import net.wurstclient.hacks.autofish.AutoFishRodSelector;
 import net.wurstclient.mixinterface.IFishingBobberEntity;
 import net.wurstclient.settings.SliderSetting;
 import net.wurstclient.settings.SliderSetting.ValueDisplay;
@@ -38,14 +36,11 @@ public final class AutoFishHack extends Hack
 			+ "Increase your range if bites are not being detected, decrease it if other people's bites are being detected as yours.",
 		1.5, 0.25, 8, 0.25, ValueDisplay.DECIMAL);
 	
-	private int bestRodValue;
-	private int bestRodSlot;
-	
 	private int castRodTimer;
 	private int reelInTimer;
-	private int scheduledWindowClick;
 	
 	private final AutoFishDebugDraw debugDraw = new AutoFishDebugDraw();
+	private final AutoFishRodSelector rodSelector = new AutoFishRodSelector();
 	
 	private boolean wasOpenWater;
 	
@@ -61,11 +56,9 @@ public final class AutoFishHack extends Hack
 	@Override
 	public void onEnable()
 	{
-		bestRodValue = -1;
-		bestRodSlot = -1;
 		castRodTimer = 0;
 		reelInTimer = -1;
-		scheduledWindowClick = -1;
+		rodSelector.reset();
 		debugDraw.reset();
 		wasOpenWater = true;
 		
@@ -91,29 +84,26 @@ public final class AutoFishHack extends Hack
 			reelInTimer--;
 		
 		ClientPlayerEntity player = MC.player;
-		PlayerInventory inventory = player.getInventory();
 		
-		if(scheduledWindowClick != -1)
+		if(rodSelector.hasScheduledClick())
 		{
-			IMC.getInteractionManager()
-				.windowClick_PICKUP(scheduledWindowClick);
-			scheduledWindowClick = -1;
+			rodSelector.doScheduledClick();
 			castRodTimer = 15;
 			return;
 		}
 		
-		updateBestRod();
+		rodSelector.updateBestRod();
 		
-		if(bestRodSlot == -1)
+		if(!rodSelector.hasARod())
 		{
 			ChatUtils.message("AutoFish has run out of fishing rods.");
 			setEnabled(false);
 			return;
 		}
 		
-		if(bestRodSlot != inventory.selectedSlot)
+		if(!rodSelector.isBestRodAlreadySelected())
 		{
-			selectBestRod();
+			rodSelector.selectBestRod();
 			return;
 		}
 		
@@ -138,78 +128,6 @@ public final class AutoFishHack extends Hack
 			reelInTimer--;
 			rightClick();
 			castRodTimer = 15;
-		}
-	}
-	
-	private void updateBestRod()
-	{
-		PlayerInventory inventory = MC.player.getInventory();
-		int selectedSlot = inventory.selectedSlot;
-		ItemStack selectedStack = inventory.getStack(selectedSlot);
-		
-		// start with selected rod
-		bestRodValue = getRodValue(selectedStack);
-		bestRodSlot = bestRodValue > -1 ? selectedSlot : -1;
-		
-		// search inventory for better rod
-		for(int slot = 0; slot < 36; slot++)
-		{
-			ItemStack stack = inventory.getStack(slot);
-			int rodValue = getRodValue(stack);
-			
-			if(rodValue > bestRodValue)
-			{
-				bestRodValue = rodValue;
-				bestRodSlot = slot;
-			}
-		}
-	}
-	
-	private int getRodValue(ItemStack stack)
-	{
-		if(stack.isEmpty() || !(stack.getItem() instanceof FishingRodItem))
-			return -1;
-		
-		int luckOTSLvl =
-			EnchantmentHelper.getLevel(Enchantments.LUCK_OF_THE_SEA, stack);
-		int lureLvl = EnchantmentHelper.getLevel(Enchantments.LURE, stack);
-		int unbreakingLvl =
-			EnchantmentHelper.getLevel(Enchantments.UNBREAKING, stack);
-		int mendingBonus =
-			EnchantmentHelper.getLevel(Enchantments.MENDING, stack);
-		int noVanishBonus = EnchantmentHelper.hasVanishingCurse(stack) ? 0 : 1;
-		
-		return luckOTSLvl * 9 + lureLvl * 9 + unbreakingLvl * 2 + mendingBonus
-			+ noVanishBonus;
-	}
-	
-	private void selectBestRod()
-	{
-		PlayerInventory inventory = MC.player.getInventory();
-		
-		if(bestRodSlot < 9)
-		{
-			inventory.selectedSlot = bestRodSlot;
-			return;
-		}
-		
-		int firstEmptySlot = inventory.getEmptySlot();
-		
-		if(firstEmptySlot != -1)
-		{
-			if(firstEmptySlot >= 9)
-				IMC.getInteractionManager()
-					.windowClick_QUICK_MOVE(36 + inventory.selectedSlot);
-			
-			IMC.getInteractionManager().windowClick_QUICK_MOVE(bestRodSlot);
-			
-		}else
-		{
-			IMC.getInteractionManager().windowClick_PICKUP(bestRodSlot);
-			IMC.getInteractionManager()
-				.windowClick_PICKUP(36 + inventory.selectedSlot);
-			
-			scheduledWindowClick = -bestRodSlot;
 		}
 	}
